Move MurmurHash to separate file due to license header, see #944
* Also removed unused 'class MurmurHash', we only use
  'object MurmurHash'
